# gen/http/openapi3.json
{"openapi":"3.0.3","info":{"title":"Concert Management API","description":"A simple API for managing music concert information","version":"1.0"},"servers":[{"url":"http://localhost:8080","description":"Concert management server"}],"paths":{"/concerts":{"get":{"tags":["concerts"],"summary":"List all concerts","description":"List concerts with optional pagination. Returns an array of concerts sorted by date.","operationId":"concerts#list","parameters":[{"name":"page","in":"query","description":"Page number","allowEmptyValue":true,"schema":{"type":"integer","description":"Page number","default":1,"example":1,"format":"int64","minimum":1},"example":1},{"name":"limit","in":"query","description":"Number of items per page","allowEmptyValue":true,"schema":{"type":"integer","description":"Number of items per page","default":10,"example":10,"format":"int64","minimum":1,"maximum":100},"example":10}],"responses":{"200":{"description":"OK response.","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Concert"},"description":"Array of concerts","example":[{"artist":"The White Stripes","date":"2024-12-25","id":"550e8400-e29b-41d4-a716-446655440000","price":7500,"venue":"Madison Square Garden, New York, NY"},{"artist":"The White Stripes","date":"2024-12-25","id":"550e8400-e29b-41d4-a716-446655440000","price":7500,"venue":"Madison Square Garden, New York, NY"},{"artist":"The White Stripes","date":"2024-12-25","id":"550e8400-e29b-41d4-a716-446655440000","price":7500,"venue":"Madison Square Garden, New York, NY"}]},"example":[{"artist":"The White Stripes","date":"2024-12-25","id":"550e8400-e29b-41d4-a716-446655440000","price":7500,"venue":"Madison Square Garden, New York, NY"},{"artist":"The White Stripes","date":"2024-12-25","id":"550e8400-e29b-41d4-a716-446655440000","price":7500,"venue":"Madison Square Garden, New York, NY"},{"artist":"The White Stripes","date":"2024-12-25","id":"550e8400-e29b-41d4-a716-446655440000","price":7500,"venue":"Madison Square Garden, New York, NY"},{"artist":"The White Stripes","date":"2024-12-25","id":"550e8400-e29b-41d4-a716-446655440000","price":7500,"venue":"Madison Square Garden, New York, NY"}]}}}}},"post":{"tags":["concerts"],"summary":"Create a new concert","description":"Create a new concert entry. All fields are required to ensure complete concert information.","operationId":"concerts#create","requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ConcertPayload"},"example":{"artist":"The White Stripes","date":"2024-12-25","price":7500,"venue":"Madison Square Garden, New York, NY"}}}},"responses":{"201":{"description":"Created response.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Concert"},"example":{"artist":"The White Stripes","date":"2024-12-25","id":"550e8400-e29b-41d4-a716-446655440000","price":7500,"venue":"Madison Square Garden, New York, NY"}}}},"400":{"description":"bad_request: Invalid input data provided","content":{"application/vnd.goa.error":{"schema":{"$ref":"#/components/schemas/Error"}}}}}}},"/concerts/{concertID}":{"delete":{"tags":["concerts"],"summary":"Delete concert","description":"Remove a concert from the system by ID. This operation cannot be undone.","operationId":"concerts#delete","parameters":[{"name":"concertID","in":"path","description":"ID of the concert to remove","required":true,"schema":{"type":"string","description":"ID of the concert to remove","example":"550e8400-e29b-41d4-a716-446655440000","format":"uuid"},"example":"550e8400-e29b-41d4-a716-446655440000"}],"responses":{"204":{"description":"No Content response."},"404":{"description":"not_found: Concert with the specified ID was not found","content":{"application/vnd.goa.error":{"schema":{"$ref":"#/components/schemas/Error"}}}}}},"get":{"tags":["concerts"],"summary":"Get concert by ID","description":"Get a single concert by its unique ID.","operationId":"concerts#show","parameters":[{"name":"concertID","in":"path","description":"Unique concert identifier","required":true,"schema":{"type":"string","description":"Unique concert identifier","example":"550e8400-e29b-41d4-a716-446655440000","format":"uuid"},"example":"550e8400-e29b-41d4-a716-446655440000"}],"responses":{"200":{"description":"OK response.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Concert"},"example":{"artist":"The White Stripes","date":"2024-12-25","id":"550e8400-e29b-41d4-a716-446655440000","price":7500,"venue":"Madison Square Garden, New York, NY"}}}},"404":{"description":"not_found: Concert with the specified ID was not found","content":{"application/vnd.goa.error":{"schema":{"$ref":"#/components/schemas/Error"}}}}}},"put":{"tags":["concerts"],"summary":"Update concert","description":"Update an existing concert by ID. Only provided fields will be updated.","operationId":"concerts#update","parameters":[{"name":"concertID","in":"path","description":"ID of the concert to update","required":true,"schema":{"type":"string","description":"ID of the concert to update","example":"550e8400-e29b-41d4-a716-446655440000","format":"uuid"},"example":"550e8400-e29b-41d4-a716-446655440000"}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateRequestBody"},"example":{"artist":"The White Stripes","date":"2024-12-25","price":7500,"venue":"Madison Square Garden, New York, NY"}}}},"responses":{"200":{"description":"OK response.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Concert"},"example":{"artist":"The White Stripes","date":"2024-12-25","id":"550e8400-e29b-41d4-a716-446655440000","price":7500,"venue":"Madison Square Garden, New York, NY"}}}},"400":{"description":"bad_request: Invalid update data provided","content":{"application/vnd.goa.error":{"schema":{"$ref":"#/components/schemas/Error"}}}},"404":{"description":"not_found: Concert with the specified ID was not found","content":{"application/vnd.goa.error":{"schema":{"$ref":"#/components/schemas/Error"}}}}}}}},"components":{"schemas":{"Concert":{"type":"object","properties":{"artist":{"type":"string","description":"The main performer for this concert","example":"The White Stripes","minLength":1,"maxLength":200},"date":{"type":"string","description":"The date when the concert will take place","example":"2024-12-25","format":"date"},"id":{"type":"string","description":"System-generated unique identifier","example":"550e8400-e29b-41d4-a716-446655440000","format":"uuid"},"price":{"type":"integer","description":"Base ticket price in cents (e.g., 7500 = $75.00)","example":7500,"format":"int64","minimum":0,"maximum":100000},"venue":{"type":"string","description":"The venue where the concert will be held","example":"Madison Square Garden, New York, NY","minLength":1,"maxLength":300}},"description":"Complete concert information including system-generated ID.","example":{"artist":"The White Stripes","date":"2024-12-25","id":"550e8400-e29b-41d4-a716-446655440000","price":7500,"venue":"Madison Square Garden, New York, NY"},"required":["id","artist","date","venue","price"]},"ConcertData":{"type":"object","properties":{"artist":{"type":"string","description":"The main performer for this concert","example":"The White Stripes","minLength":1,"maxLength":200},"date":{"type":"string","description":"The date when the concert will take place","example":"2024-12-25","format":"date"},"price":{"type":"integer","description":"Base ticket price in cents (e.g., 7500 = $75.00)","example":7500,"format":"int64","minimum":0,"maximum":100000},"venue":{"type":"string","description":"The venue where the concert will be held","example":"Madison Square Garden, New York, NY","minLength":1,"maxLength":300}},"description":"Concert information fields.","example":{"artist":"The White Stripes","date":"2024-12-25","price":7500,"venue":"Madison Square Garden, New York, NY"}},"ConcertPayload":{"type":"object","properties":{"artist":{"type":"string","description":"The main performer for this concert","example":"The White Stripes","minLength":1,"maxLength":200},"date":{"type":"string","description":"The date when the concert will take place","example":"2024-12-25","format":"date"},"price":{"type":"integer","description":"Base ticket price in cents (e.g., 7500 = $75.00)","example":7500,"format":"int64","minimum":0,"maximum":100000},"venue":{"type":"string","description":"The venue where the concert will be held","example":"Madison Square Garden, New York, NY","minLength":1,"maxLength":300}},"description":"Concert information required for creating a new concert entry.","example":{"artist":"The White Stripes","date":"2024-12-25","price":7500,"venue":"Madison Square Garden, New York, NY"},"required":["artist","date","venue","price"]},"Error":{"type":"object","properties":{"fault":{"type":"boolean","description":"Is the error a server-side fault?","example":true},"id":{"type":"string","description":"ID is a unique identifier for this particular occurrence of the problem.","example":"123abc"},"message":{"type":"string","description":"Message is a human-readable explanation specific to this occurrence of the problem.","example":"parameter 'p' must be an integer"},"name":{"type":"string","description":"Name is the name of this class of errors.","example":"bad_request"},"temporary":{"type":"boolean","description":"Is the error temporary?","example":false},"timeout":{"type":"boolean","description":"Is the error a timeout?","example":false}},"description":"Invalid input data provided","example":{"fault":true,"id":"123abc","message":"parameter 'p' must be an integer","name":"bad_request","temporary":true,"timeout":true},"required":["name","id","message","temporary","timeout","fault"]},"UpdateRequestBody":{"type":"object","properties":{"artist":{"type":"string","description":"The main performer for this concert","example":"The White Stripes","minLength":1,"maxLength":200},"date":{"type":"string","description":"The date when the concert will take place","example":"2024-12-25","format":"date"},"price":{"type":"integer","description":"Base ticket price in cents (e.g., 7500 = $75.00)","example":7500,"format":"int64","minimum":0,"maximum":100000},"venue":{"type":"string","description":"The venue where the concert will be held","example":"Madison Square Garden, New York, NY","minLength":1,"maxLength":300}},"example":{"artist":"The White Stripes","date":"2024-12-25","price":7500,"venue":"Madison Square Garden, New York, NY"}}}},"tags":[{"name":"concerts","description":"The concerts service manages music concert data. It provides CRUD operations for concert information including artist details, venues, dates, and pricing."}]}